pyproject.toml
[tool.poetry]
name = "what-s-for-dinner-bot"
version = "0.1.0"
description = "Answering the age old question via telegram!"
authors = ["Fedor Degtjarev <fed.degtjarev@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.5"
requests = "^2.32.3"
pydantic = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
